refactor:modify line and id type
id의 타입을 BigInteger에서 Long으로 수정, 줄정리

+++ src/main/java/domain/OutPlayer.java
@@ -5,15 +5,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class OutPlayer {
-    private BigInteger id;
-    private BigInteger playerId;
+    private Long id;
+    private Long playerId;
     private Reason reason;
     private LocalDateTime createdAt;
 }
