+++ src/main/java/domain/Team.java
@@ -5,15 +5,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class Team {
-    private BigInteger id;
-    private BigInteger stadiumId;
+    private Long id;
+    private Long stadiumId;
     private String name;
     private LocalDateTime CreatedAt;
 }
